+++ src/PasswordHasher.php
@@ -4,7 +4,15 @@ declare(strict_types=1);
 
 namespace Yiisoft\Security;
 
+use InvalidArgumentException;
 use SensitiveParameter;
+
+use function password_hash;
+use function password_needs_rehash;
+use function password_verify;
+
+use const PASSWORD_BCRYPT;
+use const PASSWORD_DEFAULT;
 
 /**
  * PasswordHasher allows generating password hash and verifying passwords against a hash.
@@ -27,8 +35,8 @@ final class PasswordHasher
      * @see https://www.php.net/manual/en/function.password-hash.php
      */
     public function __construct(
-        private readonly ?string $algorithm = PASSWORD_DEFAULT,
-        ?array $parameters = null,
+        private readonly string|null $algorithm = PASSWORD_DEFAULT,
+        array|null $parameters = null,
     ) {
         if ($parameters === null) {
             $this->parameters = self::SAFE_PARAMETERS[$this->algorithm] ?? [];
@@ -79,7 +87,7 @@ final class PasswordHasher
      * @param string $password The password to verify.
      * @param string $hash The hash to verify the password against.
      *
-     * @throws \InvalidArgumentException on bad password/hash parameters or if crypt() with Blowfish hash is not
+     * @throws InvalidArgumentException on bad password/hash parameters or if crypt() with Blowfish hash is not
      * available.
      *
      * @return bool whether the password is correct.
@@ -93,9 +101,25 @@ final class PasswordHasher
         string $hash
     ): bool {
         if ($password === '') {
-            throw new \InvalidArgumentException('Password must be a string and cannot be empty.');
+            throw new InvalidArgumentException('Password must be a string and cannot be empty.');
         }
 
         return password_verify($password, $hash);
+    }
+
+    /**
+     * Verifies if a hash needs rehash.
+     *
+     * @param string $hash The hash to verify.
+     *
+     * @return bool Whether rehash is needed.
+     *
+     * @see https://www.php.net/manual/function.password-needs-rehash.php
+     */
+    public function needsRehash(
+        #[SensitiveParameter]
+        string $hash
+    ): bool {
+        return password_needs_rehash($hash, $this->algorithm, $this->parameters);
     }
 }